// src/reason/layers/Block/Background/Attachment/BackgroundAttachmentStyle.css.gen.tsx
/* TypeScript file generated from BackgroundAttachmentStyle.css.res by genType. */
/* eslint-disable import/first */

// @ts-ignore: Implicit any on import
import * as BackgroundAttachmentStyle_cssBS__Es6Import from './BackgroundAttachmentStyle.css.js'
const BackgroundAttachmentStyle_cssBS: any =
  BackgroundAttachmentStyle_cssBS__Es6Import

// tslint:disable-next-line:interface-over-type-literal
export type BackgroundAttachmentStyle_t =
  | 'local'
  | 'revert'
  | 'scroll'
  | 'fixed'
  | 'initial'
  | 'inherit'
  | 'unset'

export interface IBackgroundAttachmentStyle_i {
  readonly backgroundAttachment: BackgroundAttachmentStyle_t
}

export const BackgroundAttachmentStyle_resolve: (
  t: 'fixed' | 'inherit' | 'initial' | 'local' | 'revert' | 'scroll' | 'unset'
) => string = BackgroundAttachmentStyle_cssBS.BackgroundAttachmentStyle.resolve

export const BackgroundAttachmentStyle: {
  resolve: (
    t: 'fixed' | 'inherit' | 'initial' | 'local' | 'revert' | 'scroll' | 'unset'
  ) => string
} = BackgroundAttachmentStyle_cssBS.BackgroundAttachmentStyle