// src/reason/layers/Block/Background/Clip/BackgroundClipStyle.css.gen.tsx
/* TypeScript file generated from BackgroundClipStyle.css.res by genType. */
/* eslint-disable import/first */

// @ts-ignore: Implicit any on import
import * as BackgroundClipStyle_cssBS__Es6Import from './BackgroundClipStyle.css.js'
const BackgroundClipStyle_cssBS: any = BackgroundClipStyle_cssBS__Es6Import

// tslint:disable-next-line:interface-over-type-literal
export type BackgroundClipStyle_t =
  | 'text'
  | 'contentBox'
  | 'revert'
  | 'initial'
  | 'paddingBox'
  | 'inherit'
  | 'unset'
  | 'borderBox'

export interface IBackgroundClipStyle_i {
  readonly backgroundClip: BackgroundClipStyle_t
}

export const BackgroundClipStyle_resolve: (
  t:
    | 'borderBox'
    | 'contentBox'
    | 'inherit'
    | 'initial'
    | 'paddingBox'
    | 'revert'
    | 'text'
    | 'unset'
) => string = BackgroundClipStyle_cssBS.BackgroundClipStyle.resolve

export const BackgroundClipStyle: {
  resolve: (
    t:
      | 'borderBox'
      | 'contentBox'
      | 'inherit'
      | 'initial'
      | 'paddingBox'
      | 'revert'
      | 'text'
      | 'unset'
  ) => string
} = BackgroundClipStyle_cssBS.BackgroundClipStyle