// src/reason/layers/Block/Background/Color/BackgroundColorStyle.css.gen.tsx
/* TypeScript file generated from BackgroundColorStyle.css.res by genType. */
/* eslint-disable import/first */

// @ts-ignore: Implicit any on import
import * as BackgroundColorStyle_cssBS__Es6Import from './BackgroundColorStyle.css.js'
const BackgroundColorStyle_cssBS: any = BackgroundColorStyle_cssBS__Es6Import

// tslint:disable-next-line:interface-over-type-literal
export type BackgroundColorStyle_t = 'white' | 'primary' | 'secondary' | 'black'

export interface IBackgroundColorStyle_i {
  readonly backgroundColor: BackgroundColorStyle_t
}

export const BackgroundColorStyle_resolve: (
  t: 'black' | 'primary' | 'secondary' | 'white'
) => string = BackgroundColorStyle_cssBS.BackgroundColorStyle.resolve

export const BackgroundColorStyle: {
  resolve: (t: 'black' | 'primary' | 'secondary' | 'white') => string
} = BackgroundColorStyle_cssBS.BackgroundColorStyle