// src/reason/layers/Block/Background/Opacity(--background-color)/BackgroundOpacityStyle.css.gen.tsx
/* TypeScript file generated from BackgroundOpacityStyle.css.res by genType. */
/* eslint-disable import/first */

// @ts-ignore: Implicit any on import
import * as BackgroundOpacityStyle_cssBS__Es6Import from './BackgroundOpacityStyle.css.js'
const BackgroundOpacityStyle_cssBS: any =
  BackgroundOpacityStyle_cssBS__Es6Import

// tslint:disable-next-line:interface-over-type-literal
export type BackgroundOpacityStyle_t =
  | 10
  | 80
  | 90
  | 70
  | 20
  | 0
  | 60
  | 30
  | 50
  | 100
  | 40

export interface IBackgroundOpacityStyle_i {
  readonly backgroundOpacity: BackgroundOpacityStyle_t
}

export const BackgroundOpacityStyle_resolve: (
  t: 0 | 10 | 100 | 20 | 30 | 40 | 50 | 60 | 70 | 80 | 90
) => string = BackgroundOpacityStyle_cssBS.BackgroundOpacityStyle.resolve

export const BackgroundOpacityStyle: {
  resolve: (t: 0 | 10 | 100 | 20 | 30 | 40 | 50 | 60 | 70 | 80 | 90) => string
} = BackgroundOpacityStyle_cssBS.BackgroundOpacityStyle