// src/reason/layers/Block/Background/Origin/BackgroundOriginStyle.css.gen.tsx
/* TypeScript file generated from BackgroundOriginStyle.css.res by genType. */
/* eslint-disable import/first */

// @ts-ignore: Implicit any on import
import * as BackgroundOriginStyle_cssBS__Es6Import from './BackgroundOriginStyle.css.js'
const BackgroundOriginStyle_cssBS: any = BackgroundOriginStyle_cssBS__Es6Import

// tslint:disable-next-line:interface-over-type-literal
export type BackgroundOriginStyle_t =
  | 'contentBox'
  | 'revert'
  | 'initial'
  | 'paddingBox'
  | 'inherit'
  | 'unset'
  | 'borderBox'

export interface IBackgroundOriginStyle_i {
  readonly backgroundOrigin: BackgroundOriginStyle_t
}

export const BackgroundOriginStyle_resolve: (
  t:
    | 'borderBox'
    | 'contentBox'
    | 'inherit'
    | 'initial'
    | 'paddingBox'
    | 'revert'
    | 'unset'
) => string = BackgroundOriginStyle_cssBS.BackgroundOriginStyle.resolve

export const BackgroundOriginStyle: {
  resolve: (
    t:
      | 'borderBox'
      | 'contentBox'
      | 'inherit'
      | 'initial'
      | 'paddingBox'
      | 'revert'
      | 'unset'
  ) => string
} = BackgroundOriginStyle_cssBS.BackgroundOriginStyle