// src/reason/layers/Block/Background/Position/BackgroundPositionStyle.css.gen.tsx
/* TypeScript file generated from BackgroundPositionStyle.css.res by genType. */
/* eslint-disable import/first */

// @ts-ignore: Implicit any on import
import * as BackgroundPositionStyle_cssBS__Es6Import from './BackgroundPositionStyle.css.js'
const BackgroundPositionStyle_cssBS: any =
  BackgroundPositionStyle_cssBS__Es6Import

// tslint:disable-next-line:interface-over-type-literal
export type BackgroundPositionStyle_t =
  | 'bottom'
  | 'right'
  | 'revert'
  | 'center'
  | 'initial'
  | 'inherit'
  | 'unset'
  | 'left'
  | 'top'

export interface IBackgroundPositionStyle_i {
  readonly backgroundPosition: BackgroundPositionStyle_t
}

export const BackgroundPositionStyle_resolve: (
  t:
    | 'bottom'
    | 'center'
    | 'inherit'
    | 'initial'
    | 'left'
    | 'revert'
    | 'right'
    | 'top'
    | 'unset'
) => string = BackgroundPositionStyle_cssBS.BackgroundPositionStyle.resolve

export const BackgroundPositionStyle: {
  resolve: (
    t:
      | 'bottom'
      | 'center'
      | 'inherit'
      | 'initial'
      | 'left'
      | 'revert'
      | 'right'
      | 'top'
      | 'unset'
  ) => string
} = BackgroundPositionStyle_cssBS.BackgroundPositionStyle