// src/reason/layers/Block/Background/Size/BackgroundSizeStyle.css.gen.tsx
/* TypeScript file generated from BackgroundSizeStyle.css.res by genType. */
/* eslint-disable import/first */

// @ts-ignore: Implicit any on import
import * as BackgroundSizeStyle_cssBS__Es6Import from './BackgroundSizeStyle.css.js'
const BackgroundSizeStyle_cssBS: any = BackgroundSizeStyle_cssBS__Es6Import

// tslint:disable-next-line:interface-over-type-literal
export type BackgroundSizeStyle_t =
  | 'auto'
  | 'revert'
  | 50
  | 'initial'
  | 'inherit'
  | 'unset'
  | 'contain'
  | 'cover'

export interface IBackgroundSizeStyle_i {
  readonly backgroundSize: BackgroundSizeStyle_t
}

export const BackgroundSizeStyle_resolve: (
  t:
    | 50
    | 'auto'
    | 'contain'
    | 'cover'
    | 'inherit'
    | 'initial'
    | 'revert'
    | 'unset'
) => string = BackgroundSizeStyle_cssBS.BackgroundSizeStyle.resolve

export const BackgroundSizeStyle: {
  resolve: (
    t:
      | 50
      | 'auto'
      | 'contain'
      | 'cover'
      | 'inherit'
      | 'initial'
      | 'revert'
      | 'unset'
  ) => string
} = BackgroundSizeStyle_cssBS.BackgroundSizeStyle