// src/reason/layers/Block/Border/Color/BorderColorStyle.css.gen.tsx
/* TypeScript file generated from BorderColorStyle.css.res by genType. */
/* eslint-disable import/first */

// @ts-ignore: Implicit any on import
import * as BorderColorStyle_cssBS__Es6Import from './BorderColorStyle.css.js'
const BorderColorStyle_cssBS: any = BorderColorStyle_cssBS__Es6Import

// tslint:disable-next-line:interface-over-type-literal
export type BorderColorStyle_t = 'white' | 'primary' | 'secondary' | 'black'

export interface IBorderColorStyle_i {
  readonly borderColor: BorderColorStyle_t
}

export const BorderColorStyle_resolve: (
  t: 'black' | 'primary' | 'secondary' | 'white'
) => string = BorderColorStyle_cssBS.BorderColorStyle.resolve

export const BorderColorStyle: {
  resolve: (t: 'black' | 'primary' | 'secondary' | 'white') => string
} = BorderColorStyle_cssBS.BorderColorStyle