// src/reason/layers/Block/Border/Opacity(--border-opacity)/BorderOpacityStyle.css.gen.tsx
/* TypeScript file generated from BorderOpacityStyle.css.res by genType. */
/* eslint-disable import/first */

// @ts-ignore: Implicit any on import
import * as BorderOpacityStyle_cssBS__Es6Import from './BorderOpacityStyle.css.js'
const BorderOpacityStyle_cssBS: any = BorderOpacityStyle_cssBS__Es6Import

// tslint:disable-next-line:interface-over-type-literal
export type BorderOpacityStyle_t =
  | 10
  | 80
  | 90
  | 70
  | 20
  | 0
  | 60
  | 30
  | 50
  | 100
  | 40

export interface IBorderOpacityStyle_i {
  readonly borderOpacity: BorderOpacityStyle_t
}

export const BorderOpacityStyle_resolve: (
  t: 0 | 10 | 100 | 20 | 30 | 40 | 50 | 60 | 70 | 80 | 90
) => string = BorderOpacityStyle_cssBS.BorderOpacityStyle.resolve

export const BorderOpacityStyle: {
  resolve: (t: 0 | 10 | 100 | 20 | 30 | 40 | 50 | 60 | 70 | 80 | 90) => string
} = BorderOpacityStyle_cssBS.BorderOpacityStyle