// src/reason/layers/Block/Content/Align(text-align)/ContentAlignStyle.css.gen.tsx
/* TypeScript file generated from ContentAlignStyle.css.res by genType. */
/* eslint-disable import/first */

// @ts-ignore: Implicit any on import
import * as ContentAlignStyle_cssBS__Es6Import from './ContentAlignStyle.css.js'
const ContentAlignStyle_cssBS: any = ContentAlignStyle_cssBS__Es6Import

// tslint:disable-next-line:interface-over-type-literal
export type ContentAlignStyle_t =
  | 'dotCenter'
  | 'matchParent'
  | 'justify'
  | 'right'
  | 'dot'
  | 'center'
  | 'initial'
  | 'start'
  | 'inherit'
  | 'justifyAll'
  | 'unset'
  | 'left'
  | 'end'

export interface IContentAlignStyle_i {
  readonly contentAlign: ContentAlignStyle_t
}

export const ContentAlignStyle_resolve: (
  t:
    | 'center'
    | 'dot'
    | 'dotCenter'
    | 'end'
    | 'inherit'
    | 'initial'
    | 'justify'
    | 'justifyAll'
    | 'left'
    | 'matchParent'
    | 'right'
    | 'start'
    | 'unset'
) => string = ContentAlignStyle_cssBS.ContentAlignStyle.resolve

export const ContentAlignStyle: {
  resolve: (
    t:
      | 'center'
      | 'dot'
      | 'dotCenter'
      | 'end'
      | 'inherit'
      | 'initial'
      | 'justify'
      | 'justifyAll'
      | 'left'
      | 'matchParent'
      | 'right'
      | 'start'
      | 'unset'
  ) => string
} = ContentAlignStyle_cssBS.ContentAlignStyle